src/requirements.txt
fastapi==0.74.1
uvicorn==0.17.6
sqlalchemy==1.4.27
psycopg2-binary==2.9.3
python-dotenv==0.19.2
celery==5.2.3
redis==4.1.2
Pillow==8.4.0
pytesseract==0.3.8
python-multipart==0.0.2
flower==1.0.
pytest==8.2.2
requests
httpx